Add top traffic runnable that checks endpoints fetched from ophan api/mostread
