+++ tools/amp-validation/endpoints/fixed.js
