+++ tools/amp-validation/package.json
@@ -4,7 +4,8 @@
   "description": "Utility that uses the amphtml-validator to validate AMP endpoints",
   "license": "Apache-2.0",
   "scripts": {
-      "start": "node index.js"
+      "start": "node index.js",
+      "top-traffic": "node top-traffic.js"
   },
   "repository": {
     "type": "git",
