+++ tools/amp-validation/top-traffic.js
@@ -1,0 +1,4 @@
+const fetchEndpoints = require('./endpoints/ophan');
+const run = require('./run');
+
+fetchEndpoints.then(run);
